style(coro): split coro.h into its own folder
